InNav indoor navigation
Refactored graph data structure. Combined dm.Beacon and dm.Place into one
data model, dm.Node. A physical beacon would be a node in the graph with
"physical" property; a place is a "virtual beacon", which is a node in
graph with "virtual" property; the user would be a "virtual beacon", which
is a special node in graph with "user" property.
This further extends the ability of the graph to represent more complex
indoor situations with limited number of physical nodes. Path can be
created between non-physical nodes such that even if we only have 1
physical node, the graph is still capable of representing complex indoor
map.
Converting user object into a node makes it easier to integrate
pathfinding with user's current location.