InNav indoor navigation
* Changed uniform random number generation to Gaussian random number
generation to mimic a normal distribution of measurement error. Mean is
the true value of distance, and standard deviation is deviation / 2. Given
a set of deviations of a set of true values, it assumes the measurement
deviation falls entirely on the range of +-2 * sigma.
* Added approximation of compass readings.
* Added extra step to ignore any distance change under 2.5 feet, which is
the average length of one step of human adult. This adds a little more
stability to the approximation of user location.